the options object for contentful content rendering was repeated in two components and was fairly large, so this has now been removed to its own script called options in the helper directory.
