+++ src/components/library/ContributorCard.js
@@ -2,10 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { documentToReactComponents } from '@contentful/rich-text-react-renderer';
-import { INLINES, BLOCKS } from '@contentful/rich-text-types';
 import Card from './Card';
-import ArticleCard from './ArticleCard';
 import renderContacts from '../helpers/renderContacts';
+import options from '../helpers/options';
 
 const StyledContributorCard = styled(Card)`
   flex-flow: row nowrap;
@@ -61,54 +60,25 @@ const StyledContributorCard = styled(Card)`
 
 `;
 
-const ContributorCard = ({ contributor }) => {
-  const options = {
-    renderNode: {
-      [BLOCKS.HEADING_1]: (node) => <h1>{node.content[0].value}</h1>,
-      [BLOCKS.HEADING_2]: (node) => <h2>{node.content[0].value}</h2>,
-      [BLOCKS.HEADING_3]: (node) => <h3>{node.content[0].value}</h3>,
-      [BLOCKS.PARAGRAPH]: (node, children) => <p>{children}</p>,
-      [BLOCKS.EMBEDDED_ASSET]: (node) => (
-        <img
-          src={node.data.target.fields.file.url}
-          alt={node.data.target.fields.description}
-        />
-      ),
-      [BLOCKS.EMBEDDED_ENTRY]: (node) => (
-        <ArticleCard
-          targetArticle={node}
-        />
-      ),
-      [INLINES.ASSET_HYPERLINK]: (node) => (
-        <a
-          href={node.data.target.fields.file.url}
-        >
-          {node.content[0].value}
-        </a>
-      ),
-    },
-  };
+const ContributorCard = ({ contributor }) => (
+  <StyledContributorCard>
+    <div>
+      <img
+        className="contributor-thumb"
+        src={contributor.fields.photo.fields.file.url}
+        alt={contributor.fields.photo.fields.description}
+      />
+    </div>
 
-  return (
-    <StyledContributorCard>
-      <div>
-        <img
-          className="contributor-thumb"
-          src={contributor.fields.photo.fields.file.url}
-          alt={contributor.fields.photo.fields.description}
-        />
+    <div>
+      <h2>{contributor.fields.name}</h2>
+      {documentToReactComponents(contributor.fields.bio, options)}
+      <div className="social-links">
+        {renderContacts(contributor)}
       </div>
-
-      <div>
-        <h2>{contributor.fields.name}</h2>
-        {documentToReactComponents(contributor.fields.bio, options)}
-        <div className="social-links">
-          {renderContacts(contributor)}
-        </div>
-      </div>
-    </StyledContributorCard>
-  );
-};
+    </div>
+  </StyledContributorCard>
+);
 
 ContributorCard.defaultProps = {
   contributor: {},
